+++ internal/crdschema/testdata/base.yaml
