+++ internal/crdschema/testdata/invalid.yaml
